change how example tests work around potential need for feature gate
For the int_128 feature, just assume that no feature gate is required
by tests, as there is no need for the tests to work on old nightlies.

For the try_from feature, it still needs a feature gate on current
nightly, so we cannot just remove the feature gate from the rustdoc
tests. Instead, the rustdoc tests are disabled using a bogus cfg, and
the tests are performed in mod tests.