remove reallocation in Integer::from(&int0 +/- &int1)
Previously, Integer::from(&int0 +/- &int1) would include a call to
gmp::mpz_init followed by a call to gmp::mpz_add or gmp::mpz_sub,
which could result in a reallocation. Now, the capacity required for
the addition is computed before a call to gmp::mpz_init2 so that no
reallocation needs to happen.