fix race condition in Small{Integer,Rational,Float,Complex}
The race condition was there because the Small* structs were Sync, so
parallel derefs were possible: while update_d() is updating the data
pointer atomically in one thread, another thread might be reading the
data pointer in a non-atomic way. Note that this is, to the best of my
knowledge, not an actual problem as both threads would first write the
same value atomically in the data pointer, and I don't think that
would result in an issue in any supported architecture. However, it is
still a race condition.

The fix is to use UnsafeCell<*mut limb_t> instead of
AtomicPtr<limb_t>, and mark the Small* structs as Send.