do not initialize value needlessly
For this parameter (char** end), in Rust (end: *mut *mut c_char),
mpfr_strtofr never dereferences *end, it only dereferences end and
writes to it if end itself is not null. Since we pass x.as_mut_ptr()
as end, end is obviously not null, and mpfr_strtofr writes a value
into *end, which we later assume_init.