avoid outputting exponential term if possible for {Float,Complex}
This does not affect {:E} and {:e} formatting, which is now also fixed
to print the exponent term even if the exponent is 0.

Also, zero is now output as "0" not as "0.0". Zero still never gets an
exponent term as there is no difference between say 0e0 and 0e5; 0e5
makes as much sense as inf@5 or NaN@5.